fix(review): extract duplicated _all_skill_mds() to contracts conftest.py
