+++ tests/contracts/test_issue_body_discipline.py
@@ -12,18 +12,7 @@ from __future__ import annotations
 
 import re
 
-from autoskillit.workspace.skills import bundled_skills_dir, bundled_skills_extended_dir
-
-
-def _all_skill_mds() -> list[tuple[str, str]]:
-    result = []
-    for skills_dir in (bundled_skills_dir(), bundled_skills_extended_dir()):
-        result.extend(
-            (d.name, (d / "SKILL.md").read_text())
-            for d in sorted(skills_dir.iterdir())
-            if d.is_dir() and (d / "SKILL.md").is_file()
-        )
-    return result
+from tests.contracts.conftest import _all_skill_mds
 
 
 def test_no_skill_appends_validation_summary_to_issue_body() -> None:
