+++ tests/contracts/test_ticket_body_size_ceiling.py
@@ -9,18 +9,7 @@ from __future__ import annotations
 
 import re
 
-from autoskillit.workspace.skills import bundled_skills_dir, bundled_skills_extended_dir
-
-
-def _all_skill_mds() -> list[tuple[str, str]]:
-    result = []
-    for skills_dir in (bundled_skills_dir(), bundled_skills_extended_dir()):
-        result.extend(
-            (d.name, (d / "SKILL.md").read_text())
-            for d in sorted(skills_dir.iterdir())
-            if d.is_dir() and (d / "SKILL.md").is_file()
-        )
-    return result
+from tests.contracts.conftest import _all_skill_mds
 
 
 def _has_own_filing_instructions(content: str) -> bool:
